feat: lintstaged part of husky
