+++ src/components/codestandards/markdownlint/rules/rule.ts
@@ -1,0 +1,11 @@
+import { ISerializer } from "../../../../utils";
+
+export abstract class Rule implements ISerializer {
+  serialize(): object {
+    return {
+      classNameAsPascalCase: {
+        "props-as-kebab-case": "value",
+      },
+    };
+  }
+}
