+++ src/components/githooksmanager/husky/index.ts
@@ -1,0 +1,3 @@
+export * from "./husky";
+export * from "./lintstaged";
+export * from "./lintstagedrule";
