+++ src/components/githooksmanager/husky/lintstagedrule.ts
